requirements.txt
junit-xml==1.8
PyGithub==1.43.7
